[FT/015] /controllers & refactor create project logic and change some words
