+++ src/controllers/projectController.js
@@ -1,6 +1,6 @@
 // import model
 import User from '../models/user.js'
-import Proyect from '../models/project.js'
+import Project from '../models/project.js'
 
 // import constructors
 import { ErrorCreator, ResponseCreator } from '../../utils/responseCreator.js'
@@ -10,30 +10,23 @@ import { defaultBoxes } from '../../utils/defaultBoxes.js'
 
 //  -----PRIVATE CONTROLLERS-----  //
 
-export const createProyect = async (req, res, next) => {
-  const { name, tag } = req.body
+export const createProject = async (req, res, next) => {
+  const project = {name: req.body.name, tag: req.body.tag, boxes: defaultBoxes}
 
   const user = await User.findById(req.userData.id)
 
   if (!user) {
-    next(new ErrorCreator('User not found', 404))
+   return next(new ErrorCreator('User not found', 404))
   }
 
-  Proyect.create({
-    name,
-    tag,
-    boxes: defaultBoxes
-  })
-    .then(async newProyect => {
-      // Add new proyect to user
-      user.projects.push(newProyect)
+  Project.create(project)
+    .then(async newProject => {
+      // Add new project to user
+      user.projects.push(newProject._id)
       await user.save()
-
-      const updatedUser = await User.findById(req.userData.id).populate({ path: 'projects', model: 'Project', transform: (doc) => { return doc.toObject() } })
-
-      res.send(new ResponseCreator('Proyect created Successfully', 201, { user: updatedUser }))
+      res.send(new ResponseCreator('Project created Successfully', 201, { project: newProject }))
     }).catch(err => {
-      console.error('ERROR: PROYECTCONTROLLER(CREATE)')
+      console.error('ERROR: PROJECTCONTROLLER(CREATE)')
       next(err)
     })
 }
@@ -48,18 +41,18 @@ export const getAllProjects = async (req, res, next) => {
       next(new ErrorCreator("User not found", 404));
     }
     if (!parseInt(page) || !parseInt(limit)) {
-      let projects = await Proyect.find({ _id: { $in: user.projects } });
+      let projects = await Project.find({ _id: { $in: user.projects } });
       return res.send(new ResponseCreator("page or limit is null", 200, { count: projects.length, projects }));
     }
 
-    const totalPages = Math.ceil(await Proyect.countDocuments({ _id: { $in: user.projects } }) / limit);
+    const totalPages = Math.ceil(await Project.countDocuments({ _id: { $in: user.projects } }) / limit);
 
-    let projects = await Proyect.find({ _id: { $in: user.projects } })
+    let projects = await Project.find({ _id: { $in: user.projects } })
       .limit(parseInt(limit))
       .skip((parseInt(page) - 1) * parseInt(limit));
     res.send(new ResponseCreator("success", 200, {totalPages, count: projects.length, projects }));
   } catch (err) {
-    console.error("ERROR: PROYECTCONTROLLER(getAllProjects)");
+    console.error("ERROR: PROJECTCONTROLLER(getAllProjects)");
     next(err);
   }
 };
@@ -67,13 +60,13 @@ export const getAllProjects = async (req, res, next) => {
 export const getProjectById = async (req, res, next) => {
   const { id } = req.params;
   try {
-    const project = await Proyect.findById(id).populate({model: 'Card', path: 'boxes.cards',transform: (doc,id) => doc ? doc.toObject() : id})
+    const project = await Project.findById(id).populate({model: 'Card', path: 'boxes.cards',transform: (doc,id) => doc ? doc.toObject() : id})
     if (!project) {
       return next(new ErrorCreator("Project not found", 404));
     }
     res.send(new ResponseCreator("project found successfully", 200, { project }));
   } catch (err) {
-    console.error("ERROR: PROYECTCONTROLLER(getProjectById)");
+    console.error("ERROR: PROJECTCONTROLLER(getProjectById)");
     next(err);
   }
 };
